+++ .gitignore
@@ -1,3 +1,3 @@
 node_modules/
 .env
-discord/tmp/
+misc/discord/tmp/
