+++ misc/discord/config.js
@@ -1,4 +1,3 @@
-if (process.env.NODE_ENV !== "production") require("dotenv").config();
 const path = require("path");
 const fs = require("fs");
