+++ misc/discord/index.js
@@ -1,0 +1,3 @@
+const messageBuilder = require("./messageBuilder");
+
+module.exports = messageBuilder;
