+++ misc/github/config/query.js
