+++ misc/github/fetcher/hasNextPage.js
