+++ misc/github/getter/getLeaderboard.js
@@ -3,7 +3,7 @@ const { getUser, searchPRs } = require("../fetcher");
 const { leaderboardQuery } = require("../config").query;
 const getOrgContributors = require("./getOrgContributors");
 
-const { GITHUB_CACHE_KEY, EXPIRY_TTL } = require("../../config/constant");
+const { GITHUB_CACHE_KEY, EXPIRY_TTL } = require("../../../config/constant");
 
 const getLeaderboard = (cache) => ({
   PR: () =>
