+++ misc/github/getter/getUserOrgValidContribution.js
