+++ misc/helpers/hook.js
