+++ misc/index.js
@@ -1,0 +1,5 @@
+const github = require("./github");
+const discord = require("./discord");
+const _jobs = require("./task/jobs");
+
+module.exports = { github, discord, _jobs };
