+++ router/main/index.js
@@ -5,7 +5,7 @@ const {
   getUserOrgValidContribution,
   getOrgContributors,
   getAllReposWithInfo,
-} = require("../../github");
+} = require("../../misc").github;
 const { members, repos, contributors, prCheck } = require("./opts");
 
 const routerContainer = (cachePreHandler) => (fastify, opts, done) => {
